+++ src/panglos/debug.h
@@ -65,7 +65,7 @@ extern const LUT Severity_lut[];
 
 #define _PO_PRINT(level, fmt, ...) \
         po_log(level, "%d %s %s %s +%d %s() : " fmt "\r\n", \
-                get_time(), get_task(), \
+                (int) get_time(), get_task(), \
                 lut(Severity_lut, level), \
                 __FILE__, __LINE__, __FUNCTION__, \
                 ## __VA_ARGS__ );
